Make NodeFinder return a lazy iterable.
This removes the need for the transform and filter methods, since
now Iterables.filter and Iterables.transform work just as well
without an extra pass. Since this was really the only value added
by the fancy new API, I've reverted it back to the three public
static methods (de-deprecating them), except that find returns an
Iterable instead of a List. (I've also switched the order of the
parameters to type, node instead of node, type in all three
methods). I also went and fixed all the call sites.