Update factory methods in several factory classes to support symbolic 1. first, start with the super factory class, ValueFactory; 2. and then update another abstract factory class, AggrValueFactory; 3. goes to those implementations of above abstract factory class,    like SimpleMatrixValueFactory and so on; 4. the most important modification is in IntraproceduraValueAnalysis,    by passing the symbolic value of a constant to the analysis framework,    we can start the prapagations with symbolic info; 5. finally, update files throwing errors which are because of my update of    constructors and factory methods above.
To Vineet,
I modified some of your files again, most are just update of constructors
or factory methods (which because I update constructors and factory methods
in the super class) and for the call to the constructors, I pass null as
a symbolic input to it. I think all the modification are not to understand.
By the way, you may clean up your AdvancedMatrixValue class when you are
free, some of the constructors in that file are unnecessary.