Factor out test generation into a script.
This seems nicer and simplifies build.xml a bit. Also remove
a bunch of unused tests; leftthreeaddresspass tests were deprecated
a while back and replaced with simplifications/left tests.
multireturnpass tests are weird because they're just copies of the
leftthreeaddresspass ones. A multi-return simplification is described
in Soroush's thesis but doesn't seem to be implemented. Anyway all of
these files were unused.